Commit #1 - Got as far as creating an employees page and an edit employees page without update functionality yet
Created a views folder that contains ejs files for editEmployee, employees and two other test files that I created for my own learning of the software. I created an index.js file that contains all my logic which includes a basic homepage, an employees page that lists all the employees and a edit employee page that allows a user to update employees. However, error functionality in the post method of edit employees isn't present at the moment but will be in the next commit.