Added in Department Functionality
Added in departments page which listed out all the departments. Then I added in a delete button which would check and see if the department was associated with an employee. If it was, an error would appear saying department couldn't be deleted. If it wasn't then the department would be removed from the departments table and the user would be redirected to departments page with the updated table.